+++ regex/CHANGELOG.md
@@ -8,6 +8,7 @@ This changelog summarizes major changes between TRegex versions relevant to lang
 * Added boolean matching mode (`execBoolean`), which allows omitting `RegexResult` objects in cases where capture groups are known to never be queried.
 * Added `MustAdvance` and `PythonMode` options for better Python support.
 * Support for named character escapes in Python regular expressions.
+* Limited support for \G anchors in Ruby regular expressions.
 
 ## Version 22.0.0